Adds experimental annotations; cleanup packaging
