+++ partiql-types/src/main/kotlin/org/partiql/value/PartiQL.kt
@@ -12,6 +12,7 @@
  * language governing permissions and limitations under the License.
  */
 
+// Not strictly necessary, but is explicit
 @file: JvmName("PartiQL")
 
 package org.partiql.value
@@ -76,234 +77,252 @@ import java.time.ZoneOffset
 import java.util.BitSet
 
 /**
- * TODO
+ * BOOL type value.
  *
  * @param value
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun boolValue(
     value: Boolean,
     annotations: Annotations = emptyList(),
 ): BoolValue = BoolValueImpl(value, annotations.toPersistentList())
 
 /**
- * TODO
+ * INT8 type value.
  *
  * @param value
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun int8Value(
     value: Byte,
     annotations: Annotations = emptyList(),
 ): Int8Value = Int8ValueImpl(value, annotations.toPersistentList())
 
 /**
- * TODO
+ * INT16 type value.
  *
  * @param value
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun int16Value(
     value: Short,
     annotations: Annotations = emptyList(),
 ): Int16Value = Int16ValueImpl(value, annotations.toPersistentList())
 
 /**
- * TODO
+ * INT32 type value.
  *
  * @param value
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun int32Value(
     value: Int,
     annotations: Annotations = emptyList(),
 ): Int32Value = Int32ValueImpl(value, annotations.toPersistentList())
 
 /**
- * TODO
+ * INT64 type value.
  *
  * @param value
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun int64Value(
     value: Long,
     annotations: Annotations = emptyList(),
 ): Int64Value = Int64ValueImpl(value, annotations.toPersistentList())
 
 /**
- * TODO
+ * INT type value.
  *
  * @param value
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun intValue(
     value: BigInteger,
     annotations: Annotations = emptyList(),
 ): IntValue = IntValueImpl(value, annotations.toPersistentList())
 
 /**
- * TODO
+ * DECIMAL type value.
  *
  * @param value
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun decimalValue(
     value: BigDecimal,
     annotations: Annotations = emptyList(),
 ): DecimalValue = DecimalValueImpl(value, annotations.toPersistentList())
 
 /**
- * TODO
+ * FLOAT32 type value.
  *
  * @param value
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun float32Value(
     value: Float,
     annotations: Annotations = emptyList(),
 ): Float32Value = Float32ValueImpl(value, annotations.toPersistentList())
 
 /**
- * TODO
+ * FLOAT64 type value.
  *
  * @param value
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun float64Value(
     value: Double,
     annotations: Annotations = emptyList(),
 ): Float64Value = Float64ValueImpl(value, annotations.toPersistentList())
 
 /**
- * TODO
+ * CHAR type value.
  *
  * @param value
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun charValue(
     value: Char,
     annotations: Annotations = emptyList(),
 ): CharValue = CharValueImpl(value, annotations.toPersistentList())
 
 /**
- * TODO
+ * STRING type value.
  *
  * @param value
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun stringValue(
     value: String,
     annotations: Annotations = emptyList(),
 ): StringValue = StringValueImpl(value, annotations.toPersistentList())
 
 /**
- * TODO
+ * SYMBOL type value.
  *
  * @param value
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun symbolValue(
     value: String,
     annotations: Annotations = emptyList(),
 ): SymbolValue = SymbolValueImpl(value, annotations.toPersistentList())
 
 /**
- * TODO
+ * CLOB type value.
  *
  * @param value
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun clobValue(
     value: ByteArray,
     annotations: Annotations = emptyList(),
 ): ClobValue = ClobValueImpl(value, annotations.toPersistentList())
 
 /**
- * TODO
+ * BINARY type value.
  *
  * @param value
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun binaryValue(
     value: BitSet,
     annotations: Annotations = emptyList(),
 ): BinaryValue = BinaryValueImpl(value, annotations.toPersistentList())
 
 /**
- * TODO
+ * BYTE type value.
  *
  * @param value
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun byteValue(
     value: Byte,
     annotations: Annotations = emptyList(),
 ): ByteValue = ByteValueImpl(value, annotations.toPersistentList())
 
 /**
- * TODO
+ * BLOB type value.
  *
  * @param value
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun blobValue(
     value: ByteArray,
     annotations: Annotations = emptyList(),
 ): BlobValue = BlobValueImpl(value, annotations.toPersistentList())
 
 /**
- * TODO
+ * DATE type value.
  *
  * @param value
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun dateValue(
     value: LocalDate,
     annotations: Annotations = emptyList(),
 ): DateValue = DateValueImpl(value, annotations.toPersistentList())
 
 /**
- * TODO
+ * TIME type value.
  *
  * @param value
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun timeValue(
     value: LocalTime,
     precision: Int = 0,
@@ -313,13 +332,14 @@ public fun timeValue(
 ): TimeValue = TimeValueImpl(value, precision, offset, withZone, annotations.toPersistentList())
 
 /**
- * TODO
+ * TIMESTAMP type value.
  *
  * @param value
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun timestampValue(
     value: LocalDateTime,
     precision: Int = 0,
@@ -329,20 +349,21 @@ public fun timestampValue(
 ): TimestampValue = TimestampValueImpl(value, precision, offset, withZone, annotations.toPersistentList())
 
 /**
- * TODO
+ * INTERVAL type value.
  *
  * @param value
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun intervalValue(
     value: Long,
     annotations: Annotations = emptyList(),
 ): IntervalValue = IntervalValueImpl(value, annotations.toPersistentList())
 
 /**
- * TODO
+ * BAG type value.
  *
  * @param T
  * @param elements
@@ -350,13 +371,14 @@ public fun intervalValue(
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun <T : PartiQLValue> bagValue(
     elements: List<T>,
     annotations: Annotations = emptyList(),
 ): BagValue<T> = BagValueImpl(elements.toPersistentList(), annotations.toPersistentList())
 
 /**
- * TODO
+ * LIST type value.
  *
  * @param T
  * @param elements
@@ -364,13 +386,14 @@ public fun <T : PartiQLValue> bagValue(
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun <T : PartiQLValue> listValue(
     elements: List<T>,
     annotations: Annotations = emptyList(),
 ): ListValue<T> = ListValueImpl(elements.toPersistentList(), annotations.toPersistentList())
 
 /**
- * TODO
+ * SEXP type value.
  *
  * @param T
  * @param elements
@@ -378,13 +401,14 @@ public fun <T : PartiQLValue> listValue(
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun <T : PartiQLValue> sexpValue(
     elements: List<T>,
     annotations: Annotations = emptyList(),
 ): SexpValue<T> = SexpValueImpl(elements.toPersistentList(), annotations.toPersistentList())
 
 /**
- * TODO
+ * STRUCT type value.
  *
  * @param T
  * @param fields
@@ -392,278 +416,296 @@ public fun <T : PartiQLValue> sexpValue(
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun <T : PartiQLValue> structValue(
     fields: List<Pair<String, T>>,
     annotations: Annotations = emptyList(),
 ): StructValue<T> = StructValueImpl(fields.toPersistentList(), annotations.toPersistentList())
 
-//
-// NULLABLE VALUE CONSTRUCTORS
-//
-
 /**
- * TODO
+ * NULL type value.
  *
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun nullValue(
     annotations: Annotations = emptyList(),
 ): NullValue = NullValueImpl(annotations.toPersistentList())
 
 /**
- * TODO
+ * MISSING type value.
  *
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun missingValue(
     annotations: Annotations = emptyList(),
 ): MissingValue = MissingValueImpl(annotations.toPersistentList())
 
 /**
- * TODO
+ * UNION(NULL, BOOL) type value.
  *
  * @param value
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun nullableBoolValue(
     value: Boolean? = null,
     annotations: Annotations = emptyList(),
 ): NullableBoolValue = NullableBoolValueImpl(value, annotations.toPersistentList())
 
 /**
- * TODO
+ * UNION(NULL, INT8) type value.
  *
  * @param value
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun nullableInt8Value(
     value: Byte? = null,
     annotations: Annotations = emptyList(),
 ): NullableInt8Value = NullableInt8ValueImpl(value, annotations.toPersistentList())
 
 /**
- * TODO
+ * UNION(NULL, INT8) type value.
  *
  * @param value
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun nullableInt16Value(
     value: Short? = null,
     annotations: Annotations = emptyList(),
 ): NullableInt16Value = NullableInt16ValueImpl(value, annotations.toPersistentList())
 
 /**
- * TODO
+ * UNION(NULL, INT16) type value.
  *
  * @param value
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun nullableInt32Value(
     value: Int? = null,
     annotations: Annotations = emptyList(),
 ): NullableInt32Value = NullableInt32ValueImpl(value, annotations.toPersistentList())
 
 /**
- * TODO
+ * UNION(NULL, INT64) type value.
  *
  * @param value
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun nullableInt64Value(
     value: Long? = null,
     annotations: Annotations = emptyList(),
 ): NullableInt64Value = NullableInt64ValueImpl(value, annotations.toPersistentList())
 
 /**
- * TODO
+ * UNION(NULL, INT) type value.
  *
  * @param value
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun nullableIntValue(
     value: BigInteger? = null,
     annotations: Annotations = emptyList(),
 ): NullableIntValue = NullableIntValueImpl(value, annotations.toPersistentList())
 
 /**
- * TODO
+ * UNION(NULL, DECIMAL) type value.
  *
  * @param value
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun nullableDecimalValue(
     value: BigDecimal? = null,
     annotations: Annotations = emptyList(),
 ): NullableDecimalValue = NullableDecimalValueImpl(value, annotations.toPersistentList())
 
 /**
- * TODO
+ * UNION(NULL, FLOAT32) type value.
  *
  * @param value
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun nullableFloat32Value(
     value: Float? = null,
     annotations: Annotations = emptyList(),
 ): NullableFloat32Value = NullableFloat32ValueImpl(value, annotations.toPersistentList())
 
 /**
- * TODO
+ * UNION(NULL, FLOAT64) type value.
  *
  * @param value
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun nullableFloat64Value(
     value: Double? = null,
     annotations: Annotations = emptyList(),
 ): NullableFloat64Value = NullableFloat64ValueImpl(value, annotations.toPersistentList())
 
 /**
- * TODO
+ * UNION(NULL, CHAR) type value.
  *
  * @param value
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun nullableCharValue(
     value: Char? = null,
     annotations: Annotations = emptyList(),
 ): NullableCharValue = NullableCharValueImpl(value, annotations.toPersistentList())
 
 /**
- * TODO
+ * UNION(NULL, STRING) type value.
  *
  * @param value
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun nullableStringValue(
     value: String? = null,
     annotations: Annotations = emptyList(),
 ): NullableStringValue = NullableStringValueImpl(value, annotations.toPersistentList())
 
 /**
- * TODO
+ * UNION(NULL, SYMBOL) type value.
  *
  * @param value
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun nullableSymbolValue(
     value: String? = null,
     annotations: Annotations = emptyList(),
 ): NullableSymbolValue = NullableSymbolValueImpl(value, annotations.toPersistentList())
 
 /**
- * TODO
+ * UNION(NULL, CLOB) type value.
  *
  * @param value
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun nullableClobValue(
     value: ByteArray? = null,
     annotations: Annotations = emptyList(),
 ): NullableClobValue = NullableClobValueImpl(value, annotations.toPersistentList())
 
 /**
- * TODO
+ * UNION(NULL, BINARY) type value.
  *
  * @param value
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun nullableBinaryValue(
     value: BitSet? = null,
     annotations: Annotations = emptyList(),
 ): NullableBinaryValue = NullableBinaryValueImpl(value, annotations.toPersistentList())
 
 /**
- * TODO
+ * UNION(NULL, BYTE) type value.
  *
  * @param value
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun nullableByteValue(
     value: Byte? = null,
     annotations: Annotations = emptyList(),
 ): NullableByteValue = NullableByteValueImpl(value, annotations.toPersistentList())
 
 /**
- * TODO
+ * UNION(NULL, BLOB) type value.
  *
  * @param value
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun nullableBlobValue(
     value: ByteArray? = null,
     annotations: Annotations = emptyList(),
 ): NullableBlobValue = NullableBlobValueImpl(value, annotations.toPersistentList())
 
 /**
- * TODO
+ * UNION(NULL, DATE) type value.
  *
  * @param value
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun nullableDateValue(
     value: LocalDate? = null,
     annotations: Annotations = emptyList(),
 ): NullableDateValue = NullableDateValueImpl(value, annotations.toPersistentList())
 
 /**
- * TODO
+ * UNION(NULL, TIME) type value.
  *
  * @param value
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun nullableTimeValue(
     annotations: Annotations = emptyList(),
 ): NullableTimeValue = NullableTimeValueImpl(null, 0, null, false, annotations.toPersistentList())
 
 /**
- * TODO
+ * UNION(NULL, TIME) type value.
  *
  * @param value
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun nullableTimeValue(
     value: LocalTime,
     precision: Int = 0,
@@ -673,47 +715,51 @@ public fun nullableTimeValue(
 ): NullableTimeValue = NullableTimeValueImpl(value, precision, offset, withZone, annotations.toPersistentList())
 
 /**
- * TODO
+ * UNION(NULL, TIMESTAMP) type value.
  *
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun nullableTimestampValue(
     annotations: Annotations = emptyList(),
 ): NullableTimestampValue = NullableTimestampValueImpl(null, 0, null, false, annotations.toPersistentList())
 
 /**
- * TODO
+ * UNION(NULL, TIMESTAMP) type value.
  *
  * @param value
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun nullableTimestampValue(
     value: LocalDateTime,
     precision: Int = 0,
     offset: ZoneOffset? = null,
     withZone: Boolean = false,
     annotations: Annotations = emptyList(),
-): NullableTimestampValue = NullableTimestampValueImpl(value, precision, offset, withZone, annotations.toPersistentList())
+): NullableTimestampValue =
+    NullableTimestampValueImpl(value, precision, offset, withZone, annotations.toPersistentList())
 
 /**
- * TODO
+ * UNION(NULL, TIMESTAMP) type value.
  *
  * @param value
  * @param annotations
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun nullableIntervalValue(
     value: Long? = null,
     annotations: Annotations = emptyList(),
 ): NullableIntervalValue = NullableIntervalValueImpl(value, annotations.toPersistentList())
 
 /**
- * TODO
+ * UNION(NULL, BAG) type value.
  *
  * @param T
  * @param elements
@@ -721,13 +767,14 @@ public fun nullableIntervalValue(
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun <T : PartiQLValue> nullableBagValue(
     elements: List<T>? = null,
     annotations: Annotations = emptyList(),
 ): NullableBagValue<T> = NullableBagValueImpl(elements?.toPersistentList(), annotations.toPersistentList())
 
 /**
- * TODO
+ * UNION(NULL, LIST) type value.
  *
  * @param T
  * @param elements
@@ -735,13 +782,14 @@ public fun <T : PartiQLValue> nullableBagValue(
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun <T : PartiQLValue> nullableListValue(
     elements: List<T>? = null,
     annotations: Annotations = emptyList(),
 ): NullableListValue<T> = NullableListValueImpl(elements?.toPersistentList(), annotations.toPersistentList())
 
 /**
- * TODO
+ * UNION(NULL, SEXP) type value.
  *
  * @param T
  * @param elements
@@ -749,13 +797,14 @@ public fun <T : PartiQLValue> nullableListValue(
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun <T : PartiQLValue> nullableSexpValue(
     elements: List<T>? = null,
     annotations: Annotations = emptyList(),
 ): NullableSexpValue<T> = NullableSexpValueImpl(elements?.toPersistentList(), annotations.toPersistentList())
 
 /**
- * TODO
+ * UNION(NULL, STRUCT) type value.
  *
  * @param T
  * @param fields
@@ -763,6 +812,7 @@ public fun <T : PartiQLValue> nullableSexpValue(
  * @return
  */
 @JvmOverloads
+@PartiQLValueExperimental
 public fun <T : PartiQLValue> nullableStructValue(
     fields: List<Pair<String, T>>? = null,
     annotations: Annotations = emptyList(),
