+++ partiql-types/src/main/kotlin/org/partiql/value/PartiQLValueExperimental.kt
@@ -1,0 +1,7 @@
+package org.partiql.value
+
+@RequiresOptIn(
+    message = "PartiQL Value requires explicit opt-in",
+    level = RequiresOptIn.Level.ERROR,
+)
+public annotation class PartiQLValueExperimental
