--- partiql-types/src/main/kotlin/org/partiql/value/PartiQLValueWriter.kt
@@ -1,8 +1,0 @@
-package org.partiql.value
-
-public interface PartiQLValueWriter {
-
-    public fun writeValue(value: PartiQLValue)
-
-    public fun writeValues(values: Iterator<PartiQLValue>)
-}
