+++ partiql-types/src/test/kotlin/org/partiql/value/io/PartiQLValueTextWriterTest.kt
@@ -1,12 +1,15 @@
-package org.partiql.value.impl
+@file:OptIn(PartiQLValueExperimental::class)
 
-import org.junit.jupiter.api.Assertions.assertEquals
+package org.partiql.value.io
+
+import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.parallel.Execution
 import org.junit.jupiter.api.parallel.ExecutionMode
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.MethodSource
 import org.partiql.value.Annotations
 import org.partiql.value.PartiQLValue
+import org.partiql.value.PartiQLValueExperimental
 import org.partiql.value.bagValue
 import org.partiql.value.boolValue
 import org.partiql.value.charValue
@@ -91,6 +94,7 @@ class PartiQLValueTextWriterTest {
     @Execution(ExecutionMode.CONCURRENT)
     fun testAnnotations(case: Case) = case.assert()
 
+    @OptIn(PartiQLValueExperimental::class)
     companion object {
 
         private val annotations: Annotations = listOf("x", "y")
@@ -732,6 +736,7 @@ class PartiQLValueTextWriterTest {
         )
     }
 
+    @OptIn(PartiQLValueExperimental::class)
     class Case(
         private val value: PartiQLValue,
         private val expected: String,
@@ -742,10 +747,10 @@ class PartiQLValueTextWriterTest {
             val buffer = ByteArrayOutputStream()
             val out = PrintStream(buffer)
             val writer = PartiQLValueTextWriter(out, formatted)
-            writer.writeValue(value)
+            writer.append(value)
             val actual = buffer.toString()
             // string equality
-            assertEquals(expected, actual)
+            Assertions.assertEquals(expected, actual)
         }
     }
 }
